Fix: Complete implementation.rs refactoring (#425)
Delete the 2,123-line implementation.rs monolith by moving its sole
function (analyze_synapses_with_cache_impl) into synapse.rs where it
belongs alongside the public API that calls it.

- Delete src/analysis/implementation.rs entirely
- Remove mod implementation from mod.rs
- Remove file-level #![allow(dead_code)] from synapse.rs
- Move implementation_tests ownership to synapse.rs
- Update AGENTS.md source layout

No functional changes — purely structural refactoring.
All existing tests pass unchanged.
